Rename details actions
