+++ src/redux/actions/detailsActions.js
